+++ Src/Sentry.Xamarin/Protocol/PageInfo.cs
@@ -1,8 +1,0 @@
-﻿namespace Sentry.Xamarin.Protocol
-{
-    internal class PageInfo
-    {
-        internal string Name { get; set; }
-        internal string BaseType { get; set; }
-    }
-}