+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changes
 
+* Renamed DisableXamarinFormsCache to DisableOfflineCaching. (#50) @lucas-zimerman
+* Update Sentry.NET SDK to 3.0.6. (#50) @lucas-zimerman
 * Add available RAM parameter (Android). (#46) @lucas-zimerman
 
 ## 1.0.1